\name{RFimex-package}
\alias{RFimex-package}
\alias{RFimex}
\docType{package}
\title{
Access to gridded data via Fimex
}

\description{
The RFimex package exposes some of the Fimex library functionality
to access gridded data, eventually interpolated gridded data.
}

\details{
\tabular{ll}{
Package: \tab RFimex\cr
Type: \tab Package\cr
Version: \tab 0.01\cr
Date: \tab 2012-12-03\cr
License: \tab LGPL-2 \cr
}

}

\author{
Heiko Klein
Maintainer: Heiko Klein <heiko.klein@met.no>
}
\references{
http://fimex.met.no
}

\keyword{ package }
\seealso{
}

\examples{
library(RFimex)

}
